[core] attribute: Parent's `exposed` takes precedence over description's
The `exposed` property, which determines whether the attribute is
displayed in the upper part of the node in the Graph Editor, is set
for each attribute individually in their node's description.

If an attribute has a parent (meaning it depends on a `ListAttribute` or
a `GroupAttribute`) whose `exposed` property value differs, it does not
make sense to display it separately. The attribute's `exposed` should be
aligned with its parent's.