[ui] Catch nodes that failed to be reloaded and send an error message
When reloading nodes, if some failed to be reloaded, the "Plugins reloaded"
message was still sent. We now keep track of those which failed to be
reloaded (meaning that the node descriptions were modified but there was
an actual error during the reloading process), and send an error message
listing the nodes that failed to be reloaded if there was any.